requirements.txt
--find-links https://devpi.divio.ch/divio/all/+simple/
--extra-index-url https://devpi.divio.ch/divio/all/+simple/
Django==1.5.7  # IMPORTANT!
Django-Select2==4.2.2
Fabric==1.6.0
Pillow==2.4.0
South==0.8.2
Whoosh==2.4.1
cmsplugin-filer==0.9.5
cssselect==0.8
django-reversion>=1.7,<1.8
django-admin-shortcuts==1.2.5
django-appconf==0.6
django-classy-tags==0.4
django-cms==3.0.1
django-commontranslations==0.1a1
django-compressor==1.3
django-contactform==1.1.7
django-filer==0.9.5
django-health-check==0.2
django-login-as==0.1
django-mptt==0.5.2  # IMPORTANT!
django-polymorphic==0.5.1
django-redis==3.3
django-redis-sessions==0.4.0
django-robots==0.9.2
django-sekizai==0.7
django-siteinfo==0.9.9
django-standard-form==1.1.0
djangocms-admin-style==0.2.1
djangocms-text-ckeditor==2.1.4
djangocms-link
djangocms-flash
djangocms-snippet
dploi-fabric==0.1.28
easy-thumbnails==1.2
gunicorn==18.0
hiredis==0.1.1
html5lib==0.95
lxml==3.1.0
newrelic==2.8.0.7
paramiko==1.10.0
polib==1.0.3
pycrypto==2.6
psycopg2==2.5.3
redis==2.8.0
requests==2.0.1
six==1.3.0
wsgiref==0.1.2
# celery dependencies, uncomment if project uses celery
# celery==3.0.21
# django-celery==3.0.17
# billiard==2.7.3.28
# kombu==2.5.12
# anyjson==0.3.3
# amqp==1.0.13
# celery-redis-unixsocket==0.3

# ALDRYN
aldryn-snake==0.0.2
aldryn-style==0.1.6
aldryn-grid-bootstrap==0.0.1
